requirements.txt
flask
flask-cors
python-docx
docxtpl
gspread
google-auth
num2words
